extract db implementation from interface
